float value fix

+++ app_interest/migrations/0001_initial.py
@@ -17,7 +17,7 @@ class Migration(migrations.Migration):
                 ('id', models.BigAutoField(auto_created=True, primary_key=True, serialize=False, verbose_name='ID')),
                 ('principal', models.IntegerField()),
                 ('rate', models.FloatField()),
-                ('time', models.IntegerField()),
+                ('time', models.FloatField()),
             ],
         ),
     ]
